chore(solidity/linter): cleaned and linted code

+++ toolchains/solidity/linter/core/solidhunter-lib/src/rules/miscellaneous/quotes.rs
@@ -24,7 +24,7 @@ impl RuleType for Quotes {
                                 line: line_idx,
                                 character: idx as u64,
                             },
-                            length: 1 as u64,
+                            length: 1u64,
                         },
                         message: format!("Use double quotes instead of single quote"),
                         severity: Some(self.data.severity),
